move createRateBar to bar.js
